+++ projects/shared/src/lib/services/normal-bids/normal-bid.application.service.ts
