+++ projects/shared/src/lib/services/normal-bids/normal-bid.infrastructure.service.ts
