+++ projects/shared/src/lib/services/normal-bids/normal-bid.service.ts
