+++ projects/shared/src/lib/services/renewable-bids/renewable-bid.infrastructure.service.ts
